Move pnet::packet -> pnet::old_packet.
This is in preparation for a new packet API (the interface should be similar,
though simplified). Unfortunately there's no deprecation path.
