+++ src/datalink/netmap.rs
@@ -28,8 +28,8 @@ use std::raw;
 use std::sync::Arc;
 
 use datalink::DataLinkChannelType;
-use packet::Packet;
-use packet::ethernet::{EthernetHeader, MutableEthernetHeader};
+use old_packet::Packet;
+use old_packet::ethernet::{EthernetHeader, MutableEthernetHeader};
 use util::{NetworkInterface};
 
 #[cfg(target_os = "linux")]
@@ -131,7 +131,7 @@ impl DataLinkSenderImpl {
 
     pub fn send_to(&mut self, packet: EthernetHeader, _dst: Option<NetworkInterface>)
         -> Option<IoResult<()>> {
-        use packet::MutablePacket;
+        use old_packet::MutablePacket;
         self.build_and_send(1, packet.packet().len(), &mut |&mut:mut eh: MutableEthernetHeader| {
             eh.clone_from(packet);
         })
