+++ src/datalink/winpcap.rs
@@ -17,8 +17,8 @@ use std::sync::Arc;
 
 use bindings::{bpf, winpcap};
 use datalink::{DataLinkChannelType};
-use packet::Packet;
-use packet::ethernet::{EthernetHeader, MutableEthernetHeader};
+use old_packet::Packet;
+use old_packet::ethernet::{EthernetHeader, MutableEthernetHeader};
 use util::NetworkInterface;
 
 struct WinPcapAdapter {
@@ -192,7 +192,7 @@ impl DataLinkSenderImpl {
 
     pub fn send_to(&mut self, packet: EthernetHeader, _dst: Option<NetworkInterface>)
         -> Option<IoResult<()>> {
-        use packet::MutablePacket;
+        use old_packet::MutablePacket;
         self.build_and_send(1, packet.packet().len(), |mut eh| {
             eh.clone_from(packet);
         })
